+++ activesupport/lib/active_support/core_ext/uri.rb
@@ -1,15 +1,5 @@
 # frozen_string_literal: true
 
-require "uri"
-
-module URI
-  class << self
-    def parser
-      ActiveSupport::Deprecation.warn(<<-MSG.squish)
-        URI.parser is deprecated and will be removed in Rails 7.0.
-        Use `URI::DEFAULT_PARSER` instead.
-      MSG
-      URI::DEFAULT_PARSER
-    end
-  end
-end
+ActiveSupport::Deprecation.warn(<<-MSG.squish)
+  `active_support/core_ext/uri` is deprecated and will be removed in Rails 7.1.
+MSG